Rename indra_network.py -> api.py. Fix imports.
